+++ src/api/reports.js
@@ -2,49 +2,52 @@ import request from '@/utils/request'
 import { getToken } from '@/utils/auth'
 import { baseName } from './utils'
 
-export async function changeState(state, id, authHost, token) {
+export async function changeState(reports, authHost, token) {
   return await request({
     baseURL: baseName(authHost),
-    url: `/api/pleroma/admin/reports/${id}`,
-    method: 'put',
+    url: `/api/pleroma/admin/reports`,
+    method: 'patch',
     headers: authHeaders(token),
-    data: { state }
+    data: { reports }
   })
 }
 
-export async function changeStatusScope(id, sensitive, visibility, authHost, token) {
+export async function fetchReports(filter, page, pageSize, authHost, token) {
+  const url = filter.length > 0
+    ? `/api/pleroma/admin/reports?state=${filter}&page=${page}&page_size=${pageSize}`
+    : `/api/pleroma/admin/reports?page=${page}&page_size=${pageSize}`
   return await request({
     baseURL: baseName(authHost),
-    url: `/api/pleroma/admin/statuses/${id}`,
-    method: 'put',
-    headers: authHeaders(token),
-    data: { sensitive, visibility }
-  })
-}
-
-export async function deleteStatus(id, authHost, token) {
-  return await request({
-    baseURL: baseName(authHost),
-    url: `/api/pleroma/admin/statuses/${id}`,
-    method: 'delete',
-    headers: authHeaders(token)
-  })
-}
-
-export async function fetchReports(limit, max_id, authHost, token) {
-  return await request({
-    baseURL: baseName(authHost),
-    url: `/api/pleroma/admin/reports?limit=${limit}&max_id=${max_id}`,
+    url,
     method: 'get',
     headers: authHeaders(token)
   })
 }
 
-export async function filterReports(filter, limit, max_id, authHost, token) {
+export async function fetchGroupedReports(authHost, token) {
   return await request({
     baseURL: baseName(authHost),
-    url: `/api/pleroma/admin/reports?state=${filter}&limit=${limit}&max_id=${max_id}`,
+    url: `/api/pleroma/admin/grouped_reports`,
     method: 'get',
+    headers: authHeaders(token)
+  })
+}
+
+export async function createNote(content, reportID, authHost, token) {
+  return await request({
+    baseURL: baseName(authHost),
+    url: `/api/pleroma/admin/reports/${reportID}/notes`,
+    method: `post`,
+    headers: authHeaders(token),
+    data: { content }
+  })
+}
+
+export async function deleteNote(noteID, reportID, authHost, token) {
+  return await request({
+    baseURL: baseName(authHost),
+    url: `/api/pleroma/admin/reports/${reportID}/notes/${noteID}`,
+    method: `delete`,
     headers: authHeaders(token)
   })
 }