+++ src/views/settings/components/Other.vue
@@ -14,8 +14,8 @@
 </template>
 
 <script>
-import i18n from '@/lang'
 import { mapGetters } from 'vuex'
+import i18n from '@/lang'
 import Setting from './Setting'
 
 export default {